Update Queries Validator to work on strings
